requirements.txt
certifi==2021.5.30
click==8.0.1
cycler==0.10.0
easyocr==1.4.1
Flask==2.0.1
imageio==2.9.0
imutils==0.5.4
itsdangerous==2.0.1
Jinja2==3.0.1
kiwisolver==1.3.2
MarkupSafe==2.0.1
matplotlib==3.4.3
networkx==2.6.3
numpy==1.21.2
opencv-python==4.5.3.56
opencv-python-headless==4.5.3.56
Pillow==8.2.0
pyparsing==2.4.7
pyspellchecker==0.6.2
python-bidi==0.4.2
python-dateutil==2.8.2
PyWavelets==1.1.1
PyYAML==5.4.1
scikit-image==0.18.3
scipy==1.7.1
six==1.16.0
tifffile==2021.8.30
torch==1.9.0
torchvision==0.10.0
typing-extensions==3.10.0.2
Werkzeug==2.0.1
